백준 2473 세 용액 (투포인터) 해결
이진탐색이 아닌데...? 어쨋든 투포인터 문제를 잘 구별해야 할듯

+++ src/Beakjoon/bj2143/Main.java
@@ -2,32 +2,32 @@ package Beakjoon.bj2143;
 import java.util.*;
 import java.io.*;
 public class Main {
-    static ArrayList<Integer> list1 = new ArrayList<>();
-    static ArrayList<Integer> list2 = new ArrayList<>();
-    static int[] sum1, sum2;
+    static ArrayList<Long> list1 = new ArrayList<>();
+    static ArrayList<Long> list2 = new ArrayList<>();
+    static long[] sum1, sum2;
     static int cap;
     static int n,m;
     public static void main(String[] args) throws NumberFormatException, IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         cap = Integer.parseInt(br.readLine());
         n = Integer.parseInt(br.readLine());
-        sum1 = new int[n];
+        sum1 = new long[n];
         String[] s = br.readLine().split(" ");
         for(int i=0; i<n; i++) {
             if(i > 0) {
-                sum1[i] = Integer.parseInt(s[i]) + sum1[i-1];
+                sum1[i] = Long.parseLong(s[i]) + sum1[i-1];
             }else {
-                sum1[i] = Integer.parseInt(s[i]);
+                sum1[i] = Long.parseLong(s[i]);
             }
         }
         m = Integer.parseInt(br.readLine());
-        sum2 = new int[m];
+        sum2 = new long[m];
         s = br.readLine().split(" ");
         for(int i=0; i<m; i++) {
             if(i > 0) {
-                sum2[i] = Integer.parseInt(s[i]) + sum2[i-1];
+                sum2[i] = Long.parseLong(s[i]) + sum2[i-1];
             }else {
-                sum2[i] = Integer.parseInt(s[i]);
+                sum2[i] = Long.parseLong(s[i]);
             }
         }
 
@@ -54,7 +54,7 @@ public class Main {
 
         Collections.sort(list1);
         Collections.sort(list2);
-        int ans = 0;
+        long ans = 0;
         for(int i=0; i<list1.size(); i++) {
             int start = 0, end = list2.size()-1;
             int mid=0;
